+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-02-18 10:57:23
+
+### Changed Files
+- `src/features/recruiting/components/BasicRecruitingView.tsx`
+
+### Statistics
+```
+ 1 file changed, 45 insertions(+), 20 deletions(-)
+```
+
 ## [Checkpoint] - 2026-02-18 10:10:39
 
 ### Changed Files